rename components to match naming convention
