+++ wls-gui-wahllokalsystem/src/components/ergebnisermittlung/MBW/stapelBC/BaseRowKandidatStimmenErfassen.vue
